cloud_function/data_cleaning/requirements.txt
functions-framework==3.*
pydantic==2.7.0
google-cloud-storage==2.16.0
pandas==2.2.2
python-dateutil
snowflake-connector-python